cope with running 'cvs admin' on files in the Attic
for some stupid reason, 'cvs admin' commands on files that don't
exist just fail silently.  previous attempts to run 'cvs admin -C'
on files in the attic were not working since those files are no
longer in the checked-out tree's CVS/Entries file.  to get around
this, do a 'cvs up -r1.1' on files in the attic to put them in
CVS/Entries, to be able to run 'cvs admin' on them.