properly deal with files with dead 1.1 revisions
files added on branches/imports have a dead 1.1 revision, so our
checkout of -r1.1 files to bring in these files.  when scanning each
file, record which version is the first we see that isn't dead, and
then individually checkout those revisions of those files.

also fixes an issue with weird files in the openbsd src tree like
sbin/isakmpd/pkcs.c which have no versions before 1.4 for some
reason.

with these changes, a full import and writing back of commitids on
the openbsd src tree is possible